% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{list_to_matrix_results}
\alias{list_to_matrix_results}
\title{Auxiliary reshape results function}
\usage{
list_to_matrix_results(results_list, rowname = "Test")
}
\arguments{
\item{results_list}{List with test results.}

\item{rowname}{Name for the test to be identified with. Default = "Test".}
}
\value{
Returns a matrix with the ready-to-print results.
}
\description{
Function takes the results of a test (list type) to matrix type for results printing.
}
\author{
Antón Figueroa Martínez (antfigtel@gmail.com)
}
