% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_covariance.R
\name{test_covariance}
\alias{test_covariance}
\title{Covariance multivariate tests}
\usage{
test_covariance(
  dat = NULL,
  Sigma_0 = NULL,
  Sigma_e = NULL,
  n = NULL,
  tests = c("simple"),
  verbose = TRUE
)
}
\arguments{
\item{dat}{data.frame or matrix to which apply the test. Default: \code{NULL}.}

\item{Sigma_0}{covariance matrix under null hypothesis. Mandatory for "simple" and "proportional" tests. Default:\code{NULL}}

\item{Sigma_e}{Sample covariance matrix. Default: \code{NULL}.}

\item{n}{number of observations. Just in case "dat = \code{NULL}". Default: \code{NULL}.}

\item{tests}{Vector to determine the tests to perform. Could take integer and character values according to: {1: "simple", 2: "proportional", 3: "independence"}.}

\item{verbose}{bool that enables console summary representation. Default: \code{TRUE}.}
}
\value{
Returns a list with the statistics and p-values of the corresponding tests.
}
\description{
Function that performs some contrasts related to the covariance matrix of a normal multivariate sample.
}
\author{
Antón Figueroa Martínez
}
