% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confidence_intervals_mean.R
\name{confidence_intervals_mean}
\alias{confidence_intervals_mean}
\title{Confidence intervals and simultaneous confidence intervals for the mean vector}
\usage{
confidence_intervals_mean(
  dat = NULL,
  x.bar = NULL,
  Sigma_c = NULL,
  n = NULL,
  a = NULL,
  alpha = 0.05,
  simultaneous = NULL
)
}
\arguments{
\item{dat}{matrix with the original data the test is based on. Default: \code{NULL}.}

\item{x.bar}{sample mean vector. Just in case "dat = \code{NULL}". Default: \code{NULL}.}

\item{Sigma_c}{sample mean corrected matrix. Just in case "dat = \code{NULL}". Default: \code{NULL}.}

\item{n}{number of observacions. Just in case "dat = \code{NULL}". Default: \code{NULL}.}

\item{a}{vector givin the linear combination of mu elements to which compute the confidence interval. Just in case "simultaneous = \code{NULL}". Default: \code{NULL}.}

\item{alpha}{significance level for the interval. Default: \code{0.05}.}

\item{simultaneous}{character that selects the method to be applied in the simultaneous confidence intervals computation. Options are: \code{scheffé} and \code{bonferroni}. Default: \code{NULL}.}
}
\value{
Returns a matrix with the extremes of the interval(s).
}
\description{
Function that computes both confidence intervals for summaries of the mean vector given by a*mu, and simultaneous confidence intervals given either by Scheffé or Bonferroni methods.
}
\author{
Antón Figueroa Martínez
}
