% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_compare_covariance.R
\name{test_compare_covariance}
\alias{test_compare_covariance}
\title{Covariance matrix comparison test between two populations}
\usage{
test_compare_covariance(dat_list = NULL, Sigma_c_list = NULL, n_list = NULL)
}
\arguments{
\item{dat_list}{list of matrices with the original data for each population's sample. Default: \code{NULL}.}

\item{Sigma_c_list}{list of sample mean corrected matrices for the each population. Just in case "dat_list = \code{NULL}". Default: \code{NULL}.}

\item{n_list}{list of numbers of observacions from the sample of each population. Just in case "dat_list = \code{NULL}". Default: \code{NULL}.}
}
\value{
Returns a list with the values of the statistic and p-value.
}
\description{
Function that performs the comparison test between k normal population covariance matrices.
}
\author{
Antón Figueroa Martínez
}
