% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_compare_mean.R
\name{test_compare_mean}
\alias{test_compare_mean}
\title{Mean vector comparison test between two populations}
\usage{
test_compare_mean(
  dat1 = NULL,
  dat2 = NULL,
  x1.bar = NULL,
  x2.bar = NULL,
  Sigma1_c = NULL,
  Sigma2_c = NULL,
  n1 = NULL,
  n2 = NULL
)
}
\arguments{
\item{dat1}{matrix with the original data the test is based on for the first population. Default: \code{NULL}.}

\item{dat2}{matrix with the original data the test is based on for the second population. Default: \code{NULL}.}

\item{x1.bar}{sample mean vector for the first population. Just in case "dat1 = \code{NULL}". Default: \code{NULL}.}

\item{x2.bar}{sample mean vector for the second population. Just in case "dat2 = \code{NULL}". Default: \code{NULL}.}

\item{Sigma1_c}{sample mean corrected matrix for the second population. Just in case "dat1 = \code{NULL}". Default: \code{NULL}.}

\item{Sigma2_c}{sample mean corrected matrix for the second population. Just in case "dat2 = \code{NULL}". Default: \code{NULL}.}

\item{n1}{number of observacions from the sample of the first population. Just in case "dat1 = \code{NULL}". Default: \code{NULL}.}

\item{n2}{number of observacions from the sample of the second population. Just in case "dat2 = \code{NULL}". Default: \code{NULL}.}
}
\value{
Returns a list with the values of the statistic(s) and p-value.
}
\description{
Function that performs the comparison test between two population mean vectors always when covariance matrices equal and unknown.
}
\author{
Antón Figueroa Martínez
}
