% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_mean.R
\name{test_mean}
\alias{test_mean}
\title{Mean multivariate test}
\usage{
test_mean(
  dat = NULL,
  mu0,
  cov.matrix = NULL,
  alpha = 0.05,
  n = NULL,
  x.bar = NULL
)
}
\arguments{
\item{dat}{data.frame or matriz to which apply the test. Default: \code{NULL}.}

\item{mu0}{mean vector under null hypothesis. Mandatory}

\item{cov.matrix}{population covariance matrix (if known). If not specified, substituted by the sample covariance matrix (with the corresponding adjustment in the test design). Default: \code{NULL}.}

\item{alpha}{pre-specified significance level. Default: 0.05.}

\item{n}{number of observations. Just in case "dat = \code{NULL}". Default: \code{NULL}.}

\item{x.bar}{sample mean vector. Just in case "dat = \code{NULL}". Default: \code{NULL}.}
}
\value{
Returns the value of the statistic and the p-value of the test.
}
\description{
Function that contrast a simple hypothesis over the means for multivariate normal populations when the covariance matrix is know and when it is not.
}
\examples{
# Data:
Sigma <- matrix(c(49, 35, 35, 52), nrow = 2)
x.bar <- c(102, 98)
mu0 <- c(100, 100)

# Function:
test_mean(n = 20, cov.matrix = Sigma, x.bar = x.bar, mu0 = mu0)

}
\author{
Antón Figueroa Martínez
(antfigtel@gmail.com)
}
